+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.0.7 - 2024.8.21
+
+[#19](https://github.com/spring-rs/spring-rs/pull/19)
+
+* spring-web add KnownWebError
+* jwt login example
+
 ## 0.0.0 - 2024.7.15
 
 Initial implementation of spring-boot plugin system